WebhookAuth/requirements.txt
black==21.10b0
click==7.1.2
Flask==1.1.4
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1
regex==2021.11.2
tomli==1.2.2
typing-extensions==3.10.0.2
Werkzeug==1.0.1
